Use Toolforge CDN instead of external assets
This helps to protect the privacy of users by not leaking accesses
from IP addresses outside of foundation-controlled servers. This is
recommended by
https://wikitech.wikimedia.org/wiki/Help:Toolforge/Web#External_assets
and while that's specific to ToolForge, the principals still apply to
other WMCS-hosted services including UTRS and ACC.

Toolforge itself proxies CDNJS, so anything available on CDNJS should
be available here too.

I've left the SRI hashes intact so you can confirm that there's
nothing untoward going on in the proxy chain, and that the files
served by the Toolforge CDN are identical to the ones previously
listed.

Note that I've not actually tested this, as I've not yet tried to get
the application up and running locally, so I'd appreciate it if someone
could give this a quick test before merging.