appeals: Store wiki id alongside name
Fully moving to the "wikis" table instead of the current config file is
a long process, but this patch brings us one step closer.

It adds a new "wiki_id" column to the "appeals" table, and backfills
said column using the old "wiki" column. Some uses are moved to the new
column, most (relating to permissions) are not yet. The permission table
does not yet operate on wiki ids, which is needed to do that
efficiently.

As a side effect, creating appeals for non-enabled wikis is now also
blocked on the server side. (Previously, the option was just removed on
the client side, but editing the DOM or similar tricks could be used to
submit disabled values).

ref #108