database: Don't try to shorten the user_agent field
The test sure is failing because it tries to reset the database by
undoing all of the migrations, and the user_agent field has values
longer than the previous length limit. This is fixed by removing the
user_agent truncation from the down migrations. As a bonus we can drop
2020_05_26_091658_expand_ua entirely.

Fixes: #706